refactor(checker): move mapped alias display query to diagnostics boundary
